+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
 * Add Too Many Requests error
 * Add GrantedPaymentInstrumentUpdate webhook support
+* Add ability to create a transaction from a shared nonce
 
 ## 2.3.0
 * Add AuthorizationAdjustment class and `authorizationAdjustments` to Transaction